backend/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.3
Flask-PyMongo==2.3.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
PyJWT==2.5.0
pymongo==4.2.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.4
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.1.2
zipp==3.9.0
